Move definition of linops based on functions to new module
